requirements.txt
flask
flask_cors
PyJWT
sqlalchemy
passlib
requests
psycopg2